requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
idna==3.10
instagrapi==2.1.2
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
